// frontend/src/app/shared/enums/notification-type.ts
export enum NotificationType {
    REVIEW_REPORTED = 'REVIEW_REPORTED',
    REVIEW_LIKED = 'REVIEW_LIKED',
    COMMENT_REPLIED = 'COMMENT_REPLIED',
    REVIEW_ADDED = 'REVIEW_ADDED',
    ASSET_PUBLISHED = 'ASSET_PUBLISHED',
    ASSET_UPDATED = 'ASSET_UPDATED',
    REVIEW_REMOVED = 'REVIEW_REMOVED',
}